[FEAT] Changes in the recipes mapper and interface
Changes were made to the recipes interface and the functions for mapping the data. The mapRecipe function was modified to optimize it and at the same time add a new property for the images of the ingredients. Additionally, the mapRecipeByCountry function was renamed to mapRecipeForCards as it is more coherent with this name