[FIX] Use url-parse to extract YouTube video ID
Previously, the video ID was extracted from the YouTube video URL using a regular expression. However, this caused errors when the URL was in a different format. Instead, the url-parse library is now used to extract the video ID from the URL, which is more reliable and can handle different URL formats. This commit updates the code to use url-parse for extracting the YouTube video ID